Fixes Github #12.
When filtering VCF all records were kept in memory. This leads to huge
memory consumption (~10G for 5Mb genome). Now, if *out_vcf* is specified
for *VariantSet.filter_vcf* then as soon as record in filtered, it is
written to the file. Thus reducing memory footprint, which is mostly
python stuff now.